+++ icpc-backend/src/notes/notes.service.ts
@@ -213,9 +213,52 @@ export class NotesService {
     return await this.noteRepository.save({ ...note, ...updateNoteDto });
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: string) {
     const note = await this.noteRepository.findOneBy({ id });
-    return await this.noteRepository.remove(note);
+    const userId = await this.userRepository
+      .createQueryBuilder('user')
+      .where('user.id = :userId', { userId: user })
+      .leftJoinAndSelect('user.role', 'role')
+      .getOne();
+    if (userId.role.role === 'admin') {
+      const commentBody = `${userId.name} ha eliminado el apunte con el título ${note.title}`;
+      const comment = this.commentRepository.create({
+        body: commentBody
+      });
+      const commentId = await this.commentRepository.save(comment);
+      const ticket = this.ticketRepository.create({
+        itemType: TicketType.NOTE,
+        operation: TicketOperation.DELETE,
+        status: TicketStatus.ACCEPTED,
+        originalNoteId: note,
+        commentId: commentId
+      });
+      const savedTicket = await this.ticketRepository.save(ticket);
+      if (savedTicket) {
+        return await this.noteRepository.remove(note);
+      } else {
+        throw new BadRequestException('Error al eliminar el apunte');
+      }
+    } else {
+      const commentBody = `${userId.name} ha eliminado el apunte con el título ${note.title}`;
+      const comment = this.commentRepository.create({
+        body: commentBody
+      });
+      const commentId = await this.commentRepository.save(comment);
+      const ticket = this.ticketRepository.create({
+        itemType: TicketType.NOTE,
+        operation: TicketOperation.DELETE,
+        status: TicketStatus.PENDING,
+        originalNoteId: note,
+        commentId: commentId
+      });
+      const savedTicket = await this.ticketRepository.save(ticket);
+      if (savedTicket) {
+        return await this.noteRepository.remove(note);
+      } else {
+        throw new BadRequestException('Error al eliminar el apunte');
+      }
+    }
   }
 
   async search(query: string): Promise<Note[]> {
